Fix tickets #45, #52, #55
#55 (preprocess): Fix signed/unsigned comparison in #embed limit check.
Change (long)file_size > limit to file_size > (size_t)limit so files
larger than LONG_MAX are correctly capped instead of bypassing the limit.

#52 (reflect): Assign stack offsets to reflected parameters.
jcc_ast_function_add_param now counts existing params and sets
param->offset = -(param_count + 1), matching the codegen calling
convention (bp[-1], bp[-2], etc.). Previously all reflected params
aliased bp + 0.

#45 (bytecode): Convert JMPT addresses during save/load.
JMPT embeds two addresses (table_addr and default_addr). Update
opcode_has_address/get_address_operand_index to include JMPT, and
handle default_addr (operand index 2) specially in both save and
load loops. Add generated regression tests for bytecode save/load
and JMPT loading.

Also update tests.py to recurse into test subdirs (pragma/) and
fix pragma tests to return 42 on success.

+++ src/preprocess.c
@@ -1447,7 +1447,7 @@ static Token *handle_embed_directive(JCC *vm, Token *tok,
 
     // Apply limit parameter
     size_t embed_size = file_size;
-    if (has_limit && (long)file_size > limit) {
+    if (has_limit && file_size > (size_t)limit) {
         embed_size = (size_t)limit;
     }
